Added note for how .node file is searched for
Added a note about how `node-gyp` searches for the `.node` file; how it will begin its search in the first directory it finds which has a `package.json`. This is an important feature that is surprisingly not mentioned at all within the documentation - I, myself, had trouble figuring out what was going wrong when I ran into problems regarding this silent feature.